+++ STM32CubeIDE/Application/src/wizfi360.c
@@ -6,18 +6,15 @@
  */
 
 /*********************************************************************************************/
-
-/* Includes ------------------------------------------------------------------*/
+/* Includes ---------------------------------------------------------------------------------*/
 
 #include <string.h>
-#include <stdio.h>
+
 #include "../includes/wizfi360.h"
 #include "../includes/wizfi360_required.h"
 
 /*********************************************************************************************/
-/* Required functions to be implemented by user ---------------------------------------------*/
-
-/* Private constants ----------------------------------------------------------*/
+/* Private constants ------------------------------------------------------------------------*/
 
 /**
  *	A string table holding tags which can be received from wizfi360 module via UART.
@@ -44,7 +41,8 @@ static const char *WIZFI360_TAGS[WIZFI360_NUM_TAGS] = {
 	[WIZFI360_TAG_ID_FAIL] = WIZFI360_TAG_FAIL
 };
 
-// TODO Create MQTT subscribe_topic list
+// TODO Create MQTT subscribe_topic list (id + string) (configurable by user)
+// TODO Create MQTT WIZFI360_TopicIdTypeDef (configurable by user)
 
 /*********************************************************************************************/
 
@@ -53,14 +51,10 @@ static const char *WIZFI360_TAGS[WIZFI360_NUM_TAGS] = {
 static void ResetDataStructure();
 static void ScanBufferForTags(uint8_t tmpChar, int i);
 static void ScanBufferForEcho(uint8_t tmpChar, int i);
+static void ScanBufferForMqttTopics(uint8_t tmpChar, int i);
 static void TagReceivedCallback(WIZFI360_TagIdTypeDef tagId, int length);
+static void MqttTopicReceivedCallback(uint32_t topicId, int length);
 
-
-// TODO Scan Buffer for MQTT topics
-// TODO Create MQTT Message Receive callback
-// TODO Disconnect from Broker
-// TODO Test Command
-// TODO Reset Command
 
 /*********************************************************************************************/
 
@@ -69,10 +63,7 @@ static void TagReceivedCallback(WIZFI360_TagIdTypeDef tagId, int length);
 /*
  * The wizfi360 handler
  */
-static WIZFI360_HandlerTypedef wizfi360;
-
-static ring_buffer_t rbuffer;
-
+WIZFI360_HandlerTypedef wizfi360;
 
 /*********************************************************************************************/
 
@@ -98,7 +89,9 @@ void WIZFI360_Initialize()
 void WIZFI360_Process()
 {
 	// The amount of received bytes.
-	uint32_t bytesAvailable = ring_buffer_num_items(&rbuffer);
+	uint32_t bytesAvailable = ring_buffer_num_items(&(wizfi360.UartRxBuffer));
+
+	// TODO: Don't rescan already scanned bytes
 
 	// Go through all bytes in ring buffer...
 	for (int i = 0; i < bytesAvailable; i++)
@@ -107,13 +100,13 @@ void WIZFI360_Process()
 		char tmpChar;
 
 		// Peek the next byte from ring buffer
-		ring_buffer_peek(&rbuffer, &tmpChar, i);
+		ring_buffer_peek(&(wizfi360.UartRxBuffer), &tmpChar, i);
 
 		// Check if ring buffer contains AT command echo
 		ScanBufferForEcho(tmpChar, i+1);
 
 		// Check if ring buffer contains mqtt message
-		// TODO: ScanBufferForMqqtMessages();
+		ScanBufferForMqttTopics(tmpChar, i+1);
 
 		// Check if ring buffer contains wizfi360 tags
 		ScanBufferForTags(tmpChar, i+1);
@@ -142,7 +135,6 @@ void WIZFI360_Reset()
 void WIZFI360_ResetHard()
 {
 	WIZFI360_PreResetHard();
-
 	WIZFI360_WriteResetPinLow();
 	WIZFI360_Delay(1);
 	WIZFI360_WriteResetPinHigh();
@@ -152,15 +144,6 @@ void WIZFI360_ResetHard()
 }
 
 
-/**
- * @brief	Resets the WizFi360 module using AT command
- * @note	After resetting the module we expect to receive WIZFI360_TAG_READY via UART
- * @retval	None
- */
-void WIZFI360_AT_ResetModule()
-{
-
-}
 
 
 /**
@@ -186,827 +169,22 @@ WIZFI360_WifiState WIZFI360_GetWifiState()
 }
 
 
-/**
- * @brief	Connects to an AP.
- * @note	Module must be in station mode.
- * @note	There must be no ongoing AT command.
- * @param	ssid   string parameter, the SSID of the target AP, MAX: 32 bytes
- * @param	pwd    string parameter, the password of the target AP, MAX: 64-byte ASCII.
- * @retval	None
- */
-void WIZFI360_AT_ConnectToAccessPoint(const char* ssid, const char* pwd)
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-	// If the module is not in station mode...
-	if (wizfi360.Mode != WIZFI360_MODE_STATION)
-	{
-		// TODO: Error handling
-	}
 
-	// The length of the ssid
-	const int ssidLength = strlen(ssid);
-
-	// If the SSID is too long...
-	if (ssidLength > WIZFI360_MAX_AP_SSID_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the password
-	const int pwdLength  = strlen(pwd);
-
-	// If the password is too long
-	if (pwdLength > WIZFI360_MAX_AP_PWD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the command (example: AT+CWJAP_CUR="ssid","pwd"<CR><LF>)
-	const int cmdLength =
-		strlen("AT+CWJAP_CUR=")
-		+ 5						// ssid and password are wrapped in quotation marks and separated by comma
-		+ ssidLength
-		+ pwdLength
-		+ 2;					// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_CWJAP_CUR;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+CWJAP_CUR=");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ssid);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, pwd);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-/**
- * @brief	Sets the Current WiFi mode (Station, Soft-AP or both)
- * @note	There must be no ongoing AT command.
- * @param	mode   The mode to be set.
- * @retval	None
- */
-void WIZFI360_AT_ConfigureMode(WIZFI360_ModeTypeDef mode)
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the command
-	const int cmdLength =
-		strlen("AT+CWMODE_CUR=")
-		+ 1							// the mode is one character (either '1', '2' or '3')
-		+ 2;						// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_CWMODE_CUR;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-
-	// Build the command
-
-	strcat(wizfi360.CommandBuffer, "AT+CWMODE_CUR=");
-
-	switch (mode)
-	{
-		case WIZFI360_MODE_STATION:
-		{
-			strcat(wizfi360.CommandBuffer, "1");
-			break;
-		}
-		case WIZFI360_MODE_SOFT_AP:
-		{
-			strcat(wizfi360.CommandBuffer, "2");
-			break;
-		}
-		case WIZFI360_MODE_BOTH:
-		{
-			strcat(wizfi360.CommandBuffer, "3");
-			break;
-		}
-		default:
-		{
-			// TODO: Error handling
-		}
-	}
-
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-
-/**
- * @brief	Enables/Disables DHCP for the selected mode.
- * @note	There must be no ongoing AT command.
- * @param	mode   The mode to for which DHCP should be enabled/disabled
- * @param	dhcp   The DHCP setting (enabled/disbaled)
- * @retval	None
- */
-void WIZFI360_AT_ConfigureDhcp(WIZFI360_ModeTypeDef mode, WIZFI360_DhcpModeTypeDef dhcp)
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the command
-	const int cmdLength =
-		strlen("AT+CWDHCP_CUR=")
-		+ 1							// the mode is one character (either '1', '2' or '3')
-		+ 1							// mode and dhcp are separated by comma
-		+ 1							// the dhcp setting is one character (either '0' or '1')
-		+ 2;						// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_CWDHCP_CUR;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+CWDHCP_CUR=");
-
-	switch (mode)
-	{
-		case WIZFI360_MODE_STATION:
-		{
-			// Append '1'
-			strcat(wizfi360.CommandBuffer, "1");
-			break;
-		}
-		case WIZFI360_MODE_SOFT_AP:
-		{
-			// Append '2'
-			strcat(wizfi360.CommandBuffer, "2");
-			break;
-		}
-		case WIZFI360_MODE_BOTH:
-		{
-			// Append '3'
-			strcat(wizfi360.CommandBuffer, "3");
-			break;
-		}
-		default:
-		{
-			// TODO: Error handling
-		}
-	}
-
-	// Append <comma>
-	strcat(wizfi360.CommandBuffer, ",");
-
-	switch (dhcp)
-	{
-		case WIZFI360_DHCP_DISABLE:
-		{
-			// Append '0'
-			strcat(wizfi360.CommandBuffer, "0");
-			break;
-		}
-		case WIZFI360_DHCP_ENABLE:
-		{
-			// Append '1'
-			strcat(wizfi360.CommandBuffer, "1");
-			break;
-		}
-		default:
-		{
-			// TODO: Error handling
-		}
-	}
-
-	// Append <CR><LF>
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-
-/**
- * @brief	Gets the SSL certificate of the WizFi module.
- * @note	There must be no ongoing AT command.
- * @param	None
- * @retval	None
- */
-void WIZFI360_AT_SetSSLCertificate()
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-
-	// TODO Finish setting SSL cert function
-
-	// The length of the command
-	const int cmdLength =
-		strlen("AT+CASEND=")
-		+1							// mode 0: delete cert, 1: generate cert
-		//+ strlen(CLIENT_SSL_CERT)		// length of certificate
-		+ 6;						// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_CASEND_SET;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+CASEND?");
-
-	// Append <CR><LF>
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-
-/**
- * @brief	Gets the SSL certificate of the WizFi module.
- * @note	There must be no ongoing AT command.
- * @param	None
- * @retval	None
- */
-void WIZFI360_AT_GetSSLCertificate()
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the command
-	const int cmdLength =
-		strlen("AT+CASEND?")
-		+ 2;						// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_CASEND_GET;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+CASEND?");
-
-	// Append <CR><LF>
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-
-/**
- * TODO: Check aliveTime range
- * @brief	Sets the Configuration of MQTT connection.
- * @note	This command should be set before connecting to a broker.
- * @note	There must be no ongoing AT command.
- * @param	userName   	string parameter, User Name used in the broker authentication Max: 50byte
- * @param	pwd     	string parameter, Password used in the broker authentication. Max: 50byte
- * @param	ClientID    string parameter, Client ID connected to the broker. Max: 50byte
- * @param	aliveTime   keep-alive time setting with the broker within the range of 30s~300s.
- * @retval	None
- */
-void WIZFI360_AT_ConfigureMqtt(const char* userName, const char*  pwd,
-		const char* clientId, uint16_t aliveTime )
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-	// If the module is not in station mode...
-	if (wizfi360.Mode != WIZFI360_MODE_STATION)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the user name
-	const int userNameLength = strlen(userName);
-
-	// If the userName is too long...
-	if (userNameLength > WIZFI360_MAX_MQTT_USERNAME_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the mqtt password
-	const int pwdLength  = strlen(pwd);
-
-	// If the password is too long
-	if (pwdLength > WIZFI360_MAX_MQTT_PWD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the client id
-	const int clientIdLength = strlen(clientId);
-
-	// If the clientId is too long...
-	if (clientIdLength > WIZFI360_MAX_MQTT_CLIENTID_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Build string from alive time
-	char sAliveTime[6] = {0};
-	sprintf(sAliveTime, "%d", aliveTime);
-
-
-
-	// The length of the command (example: AT+MQTTSET="user","password","id",30<CR><LF>)
-	const int cmdLength =
-		strlen("AT+MQTTSET=")
-		+ 6						// username, password and clientId are wrapped in quotation marks
-		+ 3						// and are separated by this amount of commas
-		+ userNameLength
-		+ pwdLength
-		+ clientIdLength
-		+ strlen(sAliveTime)	// the length of alive time (as string)
-		+ 2;					// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_MQTTSET;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+MQTTSET=");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, userName);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, pwd);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, clientId);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, sAliveTime);
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-
-/**
- * TODO: NULL check
- * TODO: (strlen < 1) check
- * TODO: Comment subTopic2 / subtopic3
- * @brief	Sets the Topic of Publish and Subscribe
- * @note	This command should be set before connecting to a broker.
- * @note	There must be no ongoing AT command.
- * @param	pubTopic	string parameter, The topic published on the WizFi360
- * @param	subscribeTopic 	string parameter, The topic subscribed by the WizFi360
- * @retval	None
- */
-void WIZFI360_AT_MqttSetTopic(const char* pubTopic, const char*  subTopic1,
-		const char* subTopic2, const char* subTopic3)
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-	// If the module is not in station mode...
-	if (wizfi360.Mode != WIZFI360_MODE_STATION)
-	{
-		// TODO: Error handling
-	}
-
-	// The length of the command (example: AT+MQTTTOPIC="pubTopic","subTopic1","SubTopic2","SubTopic3"<CR><LF>)
-	const int cmdLength =
-		strlen("AT+MQTTTOPIC=")
-		+ 2						// publishTopic is wrapped in quotation marks
-		+ 2						// each subscribeTopic is wrapped in quotation marks
-		+ 1						// topics are separated by this amount of commas
-		+ strlen(pubTopic)
-		+ strlen(subTopic1)
-		+ 2;					// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_MQTTTOPIC;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+MQTTTOPIC=");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, pubTopic);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, subTopic1);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-
-/**
- * @brief	Connects to a Broker
- * @note	There must be no ongoing AT command.
- * @note	Whenever messages of subscribe topic is received, it will return as below:
- * 			<subscribe topic> -> "subscribe data"
- * @param	authMode		Decides, whether to connect to a broker with/without authentication
- * @param	mqttBrokerIP	string parameter indicating the broker IP address
- * @param	mqttBrokerPort 	the broker port number
- * @retval	None
- */
-void WIZFI360_AT_MqttConnectToBroker(WIZFI360_MqqtAuthModeTypeDef authMode,
-		const char*  mqttBrokerIP, uint16_t mqttBrokerPort)
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-	// If the module is not in station mode...
-	if (wizfi360.Mode != WIZFI360_MODE_STATION)
-	{
-		// TODO: Error handling
-	}
-
-	// Build string from alive time
-	char sPort[6] = {0};
-	sprintf(sPort, "%d", mqttBrokerPort);
-
-	// The length of the command (example: AT+MQTTCON=0,"35.156.215.0",1883<CR><LF>)
-	const int cmdLength =
-		strlen("AT+MQTTCON=")
-		+ 2						// IP address is wrapped in quotation marks
-		+ 2						// the parameters are separated by this amount of commas
-		+ 1						// authMode is one character long ('0' or '1')
-		+ strlen(sPort)
-		+ strlen(mqttBrokerIP)
-		+ 2;					// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_MQTTCON;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+MQTTCON=");
-
-	switch (authMode)
-	{
-		case WIZFI360_MQTT_AUTH_DISABLE:
-		{
-			strcat(wizfi360.CommandBuffer, "0");
-			break;
-		}
-		case WIZFI360_MQQT_AUTH_ENABLE:
-		{
-			strcat(wizfi360.CommandBuffer, "1");
-			break;
-		}
-		default:
-		{
-			// TODO: Error handling
-		}
-	}
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, mqttBrokerIP);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, ",");
-	strcat(wizfi360.CommandBuffer, sPort);
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
-
-/**
- * @brief	Connects to a Broker TODO COMMENT
- * @note	There must be no ongoing AT command.
- * @note	Whenever messages of subscribe topic is received, it will return as below:
- * 			<subscribe topic> -> "subscribe data"
- * @param	authMode		Decides, whether to connect to a broker with/without authentication
- * @param	mqttBrokerIP	string parameter indicating the broker IP address
- * @param	mqttBrokerPort 	the broker port number
- * @retval	None
- */
-void WIZFI360_AT_MqttPublishMessage(const char* message)
-{
-	// If there is an ongoing AT command...
-	if (wizfi360.ExpectingResponse)
-	{
-		// TODO: Error handling
-	}
-	// TODO if is not connected to broker
-	// If the module is not in station mode...
-	if (wizfi360.Mode != WIZFI360_MODE_STATION)
-	{
-		// TODO: Error handling
-	}
-
-
-	// The length of the command (example: AT+MQTTCON=0,"35.156.215.0",1883<CR><LF>)
-	const int cmdLength =
-		strlen("AT+MQTTPUB=")
-		+ 2						// the message is wrapped in quotation marks
-		+ strlen(message)
-		+ 2;					// command ends with <CR><LF>
-
-	// If the command is too long...
-	if (cmdLength >= WIZFI360_MAX_CMD_LEN)
-	{
-		// TODO: Error handling
-	}
-
-	// Write the command id into wizfi360 structure
-	wizfi360.CommandId = WIZFI360_CMD_ID_MQTTPUB;
-
-	// Write the command length into wizfi360 structure
-	wizfi360.CommandLength = cmdLength;
-
-	// Empty the command buffer string
-	wizfi360.CommandBuffer[0] = '\0';
-
-	// Build the command
-	strcat(wizfi360.CommandBuffer, "AT+MQTTPUB=");
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, message);
-	strcat(wizfi360.CommandBuffer, "\"");
-	strcat(wizfi360.CommandBuffer, "\r\n");
-
-#ifdef WIZFI360_UART_TX_MODE_NON_BLOCKING
-	// Send the command
-	WIZFI360_UART_SendNonBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength);
-#else
-	// Send the command
-	WIZFI360_UART_SendBlockingMode((uint8_t*) wizfi360.CommandBuffer,
-			wizfi360.CommandLength, 10000);
-#endif
-
-	// We expect a response for this command.
-	wizfi360.ExpectingResponse = 1;
-
-	// If echo mode is enabled...
-	if (wizfi360.EchoEnabled)
-	{
-		// We expect an echo for this command
-		wizfi360.ExpectingEcho = 1;
-	}
-}
 
 /*
- * TODO: Comment WIZFI360_UART_BytesReceived
+ * TODO: Comment on WIZFI360_UART_BytesReceived
  */
 void WIZFI360_UART_BytesReceived(const char *data, ring_buffer_size_t size)
 {
-	ring_buffer_queue_arr(&rbuffer, data, size);
+	ring_buffer_queue_arr(&(wizfi360.UartRxBuffer), data, size);
 }
 
 /*
- * TODO: Comment WIZFI360_UART_ByteReceived
+ * TODO: Comment on WIZFI360_UART_ByteReceived
  */
 void WIZFI360_UART_ByteReceived(const char data)
 {
-	ring_buffer_queue(&rbuffer, data);
+	ring_buffer_queue(&(wizfi360.UartRxBuffer), data);
 }
 
 /*********************************************************************************************/
@@ -1015,13 +193,30 @@ void WIZFI360_UART_ByteReceived(const char data)
 static void ResetDataStructure()
 {
 	/* Initialize ring buffer */
-	ring_buffer_init(&rbuffer);
+	ring_buffer_init(&(wizfi360.UartRxBuffer));
 
-	// TODO: DO PROPER RESET
-	wizfi360.EchoEnabled = 1;
+	wizfi360.WifiMode = WIZFI360_MODE_STATION;
+
+	wizfi360.EchoMode = WIZFI360_ECHO_ENABLE;
+
 	wizfi360.ExpectingResponse = 0;
-	wizfi360.Mode = WIZFI360_MODE_STATION;
+
+	wizfi360.ExpectingEcho = 0;
+
 	wizfi360.WifiState = WIZFI360_WIFI_DISCONNECTED;
+
+	wizfi360.CommandBuffer[0] = '\0';
+
+	wizfi360.CommandLength = 0;
+
+	wizfi360.CommandId = -1;
+
+	wizfi360.EchoCharsReceived = 0;
+
+	for (int tagId = 0; tagId < WIZFI360_NUM_TAGS; tagId++)
+	{
+		wizfi360.TagCharsReceived[tagId] = 0;
+	}
 }
 
 
@@ -1053,7 +248,7 @@ static void ScanBufferForEcho(uint8_t tmpChar, int i)
 			// A buffer to read the echo into (the length of the echo equals command length)
 			uint8_t echo[wizfi360.CommandLength];
 			// Read the echo from UART buffer
-			ring_buffer_dequeue_arr(&rbuffer, (char*)echo, wizfi360.CommandLength);
+			ring_buffer_dequeue_arr(&(wizfi360.UartRxBuffer), (char*)echo, wizfi360.CommandLength);
 			// We do not expect an echo anymore
 			wizfi360.ExpectingEcho = 0;
 			// Reset the counter.
@@ -1118,7 +313,20 @@ static void ScanBufferForTags(uint8_t tmpChar, int i)
 	}
 }
 
-/*********************************************************************************************/
+
+/**
+ * TODO: Comment on ScanBufferForMqttTopics
+ */
+static void ScanBufferForMqttTopics(uint8_t tmpChar, int i)
+{
+	// TODO: Implement ScanBufferForMqttTopics
+
+	// TODO: Call MqttTopicReceivedCallback, if topic found
+	if (1)
+	{
+		MqttTopicReceivedCallback(0,0);
+	}
+}
 
 
 /**
@@ -1132,60 +340,22 @@ static void ScanBufferForTags(uint8_t tmpChar, int i)
 static void TagReceivedCallback(WIZFI360_TagIdTypeDef tagId, int length)
 {
 	uint8_t data[length];
-	ring_buffer_dequeue_arr(&rbuffer, (char*)data, length);
+	ring_buffer_dequeue_arr(&(wizfi360.UartRxBuffer), (char*)data, length);
 
 	switch(tagId)
 	{
-		// If we received the OK tag...
+		// If we received an AT Response...
 		case WIZFI360_TAG_ID_OK:
-		{
-			wizfi360.ExpectingResponse = 0;
-
-			#ifdef WIZFI360_CALLBACK_USED_COMMAND_COMPLETE
-			WIZFI360_CommandCpltCallback(wizfi360.CommandId,
-					WIZFI360_RESPONSE_OK);
-			#endif
-
-			break;
-		}
-		case WIZFI360_TAG_ID_SEND_OK:
-		{
-			wizfi360.ExpectingResponse = 0;
-			break;
-		}
 		case WIZFI360_TAG_ID_ERROR:
-		{
-			wizfi360.ExpectingResponse = 0;
-
-			#ifdef WIZFI360_CALLBACK_USED_COMMAND_COMPLETE
-			WIZFI360_CommandCpltCallback(wizfi360.CommandId,
-					WIZFI360_RESPONSE_ERROR);
-			#endif
-
-			break;
-		}
-		case WIZFI360_TAG_ID_ALREADY_CONNECTED:
-		{
-			wizfi360.ExpectingResponse = 0;
-			break;
-		}
-		case WIZFI360_TAG_ID_SEND_FAIL:
-		{
-			wizfi360.ExpectingResponse = 0;
-			break;
-		}
 		case WIZFI360_TAG_ID_FAIL:
+		case WIZFI360_TAG_ID_SEND_OK:
+		case WIZFI360_TAG_ID_SEND_FAIL:
+		case WIZFI360_TAG_ID_ALREADY_CONNECTED:
+		case WIZFI360_TAG_ID_BUSY_SENDING:
+		case WIZFI360_TAG_ID_BUSY_PROCESSING:
 		{
-			#ifdef WIZFI360_CALLBACK_USED_WIFI_CONNECT_FAILED
-			WIZFI360_WifiConnectFailedCallback();
-			#endif
-
-			if (wizfi360.CommandId == WIZFI360_CMD_ID_CWJAP_CUR &&
-					wizfi360.ExpectingResponse)
-			{
-				wizfi360.ExpectingResponse = 0;
-			}
-
+			// Handle the response.
+			WIZFI360_AT_HandleResponse(tagId);
 			break;
 		}
 		case WIZFI360_TAG_ID_READY:
@@ -1218,19 +388,20 @@ static void TagReceivedCallback(WIZFI360_TagIdTypeDef tagId, int length)
 			#endif
 			break;
 		}
-		case WIZFI360_TAG_ID_BUSY_SENDING:
-		{
-			break;
-		}
-		case WIZFI360_TAG_ID_BUSY_PROCESSING:
-		{
-			break;
-		}
 		default:
 		{
 			break;
 		}
 	}
+}
+
+/**
+ * TODO: Comment on MqttTopicReceivedCallback
+ */
+static void MqttTopicReceivedCallback(uint32_t topicId, int length)
+{
+	// TODO: Implement MqttTopicReceivedCallback
+	// TODO: Notify User about received MqttTopic
 }
 
 /*********************************************************************************************/
